Add product service

+++ Entities/Models/Product.cs
@@ -11,15 +11,15 @@ public class Product
     public string ProductName { get; set; }
     [Required(ErrorMessage = "Calories is required field.")]
     [Range(1, Int32.MaxValue)]
-    public int Calories { get; set; }  
+    public double Calories { get; set; }  
     [MaxLength(256,ErrorMessage = "Maximum length for the Description is 256 characters.")]
     public string? Description { get; set; }
     [MaxLength(64,ErrorMessage = "Maximum length for the Brand is 64 characters.")]
     public string? Brand { get; set; }
-    public int? ServingSize { get; set; }
-    public int? Protein { get; set; }
-    public int? Fat { get; set; }
-    public int? Carbs { get; set; }
+    public double? ServingSize { get; set; }
+    public double? Protein { get; set; }
+    public double? Fat { get; set; }
+    public double? Carbs { get; set; }
     public string? Image { get; set; }
     public string UserId { get; set; }
 }
